Adjust versioned build-deps.
For the time being, downgrade them to the versions that are immediately
available in bullseye, which means we have build dependencies that can
be satisfied at build time, but also dependencies that can be satisfied
at run time.

If those versions aren't enough, we might need to look into backporting
those two packages.